Step 1 of reducing `CURRENT_TAG`s
Pulling the value of a reference could be a side-effectful operation
from the perspective of the revision timeline (e.g. updating an
`UpdatableTag` is a change on the timeline). Therefore we should always
compute the value first, then pull from the validation tag.

If we always follow that protocol, there is no actual reason to
initialize an `UpdatableTag` with a `CURRENT_TAG`. While this is always
"correct", it causes values that doesn't otherwise matter to "poison" an
tag combinator all the way up. (And also causes caches to miss – because
the tag value might have changed after computing the value, as it is
usually the case when an `UpdatableTag` is involved.)