[GLIMMER2] Optimize {{property}} PropertyReference
A quick survey in our app indicates ~77% of `PropertyReference` have a
const parent. This is unsurprising, because most path lookup curlies are
single-level (no dots in the path) lookup on the component directly.

This means we can do less work and book-keeping in those cases (checking
if the parent is a proxy on every `compute`, etc).