Call IsVisibleBinding only if no style binding
Previously, the `seen` variable was used as both an `Array` and a
hashmap (plain JavaScript object) when it was really an `Array`. The
`Array` would never return `true` for `seen['style']` in this code block
meaning that the `IsVisibleBinding` call would always be invoked,
regardless of whether there was an attribute binding for `style`. This
behavior is most likely a bug. The author intends to investigate unit /
integration tests for this case and add them in a separate commit.