[GLIMMER2] Component instrumentation support
Added support for the following instrumentation events:

- render.component
- render.outlet (new)

Previously, all outlets are backed by an “OutletView” instance. Since we
don’t do that anymore, I have to replace that with a new event to make
the test pass.

The `render.component` event also gets a new `initialRender` field in
the payload that indicates whether this is an initial render or a
re-render.