😪 Make `view.childViews` lazy
🆕 `getRootViews` return an array of “top-level” components
🆕 `getChildViews` to replace `view.childViews` (to be deprecated 🔜)
🆕 “tagless” components are registered in the view-registry with a GUID
🆒 `view.childViews` is now a getter that delegates to `getChildViews`
🆓 `view.childViews` no longer leak destroyed components (bug in beta)
🔙 `view.childViews` now tracks updates (components added or removed
   during updates) correctly, which was unreliable since 1.13

Instead of tracking `childView` instances, we now track a list of IDs
that can be used to look them up in the view registry. This has the
benefit of not holding a reference to the view instance – thus do not
require cleanup when the child views are destroyed.

When requested, we lazily query the view registry to reify the (still)
live children and cleanup and stale IDs while we are at it.

This approach does leak the string IDs of destroyed components, however
the effect is believed to be minimal. If this ended up causing problems,
we can be smarter about this, such as tracking the number of removals
and schedule a collection (outside of the render loop) when a view has
accumulated enough removals.