Temporary fix for `{{action` usage in Fastboot.
This is a temporary fix to prevent errors when using `<div {{action
'foo'}}></div>` in FastBoot.

The longer term fix would be to implement the concept of a
`ModifierManager` (roughly in the same vain as `ComponentManager`) that
has hooks and can make the determination of when to fire `install` /
`update` on the underlying modifier.