[BUGFIX release] Fix component action bubbling semantics.
When a component action returns `true` we should not bubble the action
upwards (this is a significant difference between `Component#send` and
`Controller#send` / `Route#send`). The only way to get bubbling from
within a component is to call `Component#sendAction`.

This removes an invalid fallback in `ActionSupport` mixin that allowed
all component actions to bubble, and adds a few tests to help ensure we
do not regress in future refactors.