[BUGFIX beta] Allow `layout` to be injected.
In prior versions of Ember templates were not instantiated, and it was
common to inject the layout into the component before looking it
up (this was actually what the framework itself did until relatively
recently). There are a number of addons (as well as component unit tests
in ember-test-helpers) that rely on this behavior.

This change brings back the ability for the template cache system to
handle `layout` as either a factory (commonly done by addons when
importing a template and adding it as a layout property) or
instance (when using registry injection).