[BUGFIX beta] Remove empty string overrides in Ember.get.
Previously, to work around bugs internal to Glimmer we made `Ember.get({
'': 'foo' }, '')` return the object itself, instead of `'foo'`. This was
because `{{this}}` in Glimmer/Handlebars is actually representated as
`''` (which I agree is a tad O_o).

Those issues have been addressed, so this override / work-around is no
longer needed.