[CLEANUP beta] Remove `Ember.View` deadcode
The deprecated `Ember.View` has been removed a while ago. The exports
in this file is not used anywhere inside the `ember-views` package and
also not re-exported from the package index. Contrary to the outdated
documentation, we don’t export the `Ember.View` global also, so this is
just dead code.

I kept the file around for now until we finish migrating the
documentation.